Spring security complete

+++ src/main/java/com/project/UrbanNest/service/imp/HotelServiceImpl.java
@@ -1,21 +1,25 @@
-package com.project.UrbanNest.service;
+package com.project.UrbanNest.service.imp;
 
 import com.project.UrbanNest.dto.HotelDto;
 import com.project.UrbanNest.dto.HotelInfoDto;
 import com.project.UrbanNest.dto.RoomDto;
 import com.project.UrbanNest.entity.Hotel;
 import com.project.UrbanNest.entity.Room;
+import com.project.UrbanNest.entity.User;
 import com.project.UrbanNest.exception.ResourceNotFoundException;
+import com.project.UrbanNest.exception.UnAuthorisedException;
 import com.project.UrbanNest.repository.HotelRepository;
 import com.project.UrbanNest.repository.RoomRepository;
+import com.project.UrbanNest.service.HotelService;
+import com.project.UrbanNest.service.InventoryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @Slf4j
@@ -32,16 +36,26 @@ public class HotelServiceImpl implements HotelService {
         log.info("Creating a new hotel with name: {}",hotelDto.getName());
         Hotel hotel =modelMapper.map(hotelDto, Hotel.class);
         hotel.setActive(false);
+
+        User user= getCurrentUser();
+        hotel.setOwner(user);
+
         Hotel savedHotel=hotelRepository.save(hotel);
         log.info("Created a new hotel with ID: {}",savedHotel.getId());
         return modelMapper.map(hotel,HotelDto.class);
     }
 
     @Override
-    public HotelDto findHotelById(Long Id) {
+    public HotelDto getHotelById(Long Id) {
         log.info("Getting the hotel with ID: {}",Id);
         Hotel hotel=hotelRepository.findById(Id)
                 .orElseThrow(()-> new ResourceNotFoundException("Hotel not found with ID: "+Id));
+
+        User user=getCurrentUser();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+user.getId());
+        }
+
         return modelMapper.map(hotel,HotelDto.class);
     }
 
@@ -50,6 +64,13 @@ public class HotelServiceImpl implements HotelService {
         log.info("Updating the hotel with ID: {}", id);
         Hotel hotel=hotelRepository.findById(id)
                 .orElseThrow(()-> new ResourceNotFoundException("Hotel not found with ID: "+id));
+
+        User user=getCurrentUser();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+user.getId());
+        }
+
+
         modelMapper.map(hotelDto,hotel);
         hotel.setId(id);
         hotel=hotelRepository.save(hotel);
@@ -59,7 +80,13 @@ public class HotelServiceImpl implements HotelService {
     @Override
     @Transactional
     public void deleteHotelById(Long id) {
-        Hotel hotel=getHotelById(id);
+        Hotel hotel=findHotelById(id);
+
+        User user=getCurrentUser();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+user.getId());
+        }
+
         for(Room room:hotel.getRooms()){
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
@@ -71,7 +98,13 @@ public class HotelServiceImpl implements HotelService {
     @Transactional
     public void activeHotel(Long hotelId) {
         log.info("Activating the hotel with ID: {}",hotelId);
-        Hotel hotel=getHotelById(hotelId);
+        Hotel hotel=findHotelById(hotelId);
+
+        User user=getCurrentUser();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+user.getId());
+        }
+
         hotel.setActive(true);
         log.info("Saving The hotel With Activation");
         hotelRepository.save(hotel);
@@ -85,7 +118,7 @@ public class HotelServiceImpl implements HotelService {
 
     @Override
     public HotelInfoDto getHotelInfoById(Long hotelId) {
-        Hotel hotel=getHotelById(hotelId);
+        Hotel hotel=findHotelById(hotelId);
         List<RoomDto> rooms= hotel.getRooms()
                 .stream()
                 .map((element) -> modelMapper.map(element, RoomDto.class))
@@ -93,9 +126,13 @@ public class HotelServiceImpl implements HotelService {
         return new HotelInfoDto(modelMapper.map(hotel,HotelDto.class),rooms);
     }
 
-    private Hotel getHotelById(Long id){
+    private Hotel findHotelById(Long id){
         return hotelRepository.findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with ID="+id));
+    }
+
+    private User getCurrentUser(){
+        return (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
     }
 
 }
