+++ src/main/java/com/project/UrbanNest/service/imp/RoomServiceImpl.java
@@ -1,14 +1,19 @@
-package com.project.UrbanNest.service;
+package com.project.UrbanNest.service.imp;
 
 import com.project.UrbanNest.dto.RoomDto;
 import com.project.UrbanNest.entity.Hotel;
 import com.project.UrbanNest.entity.Room;
+import com.project.UrbanNest.entity.User;
 import com.project.UrbanNest.exception.ResourceNotFoundException;
+import com.project.UrbanNest.exception.UnAuthorisedException;
 import com.project.UrbanNest.repository.HotelRepository;
 import com.project.UrbanNest.repository.RoomRepository;
+import com.project.UrbanNest.service.InventoryService;
+import com.project.UrbanNest.service.RoomService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,7 +24,7 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class RoomServiceImpl implements RoomService{
+public class RoomServiceImpl implements RoomService {
 
     private final RoomRepository roomRepository;
     private final HotelRepository hotelRepository;
@@ -30,6 +35,12 @@ public class RoomServiceImpl implements RoomService{
     public RoomDto createNewRoom(Long hotelId,RoomDto roomDto) {
         log.info("Creating a new room in hotel with ID:{}",hotelId);
         Hotel hotel=getHotelById(hotelId);
+
+        User user=getCurrentUser();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
+
         Room room=modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
         room= roomRepository.save(room);
@@ -45,6 +56,12 @@ public class RoomServiceImpl implements RoomService{
     public List<RoomDto> getAllRoomInHotel(Long hotelId) {
         log.info("Getting all rooms in hotel with ID: {}",hotelId);
         Hotel hotel=getHotelById(hotelId);
+
+        User user=getCurrentUser();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotel.getId());
+        }
+
         return hotel.getRooms()
                 .stream()
                 .map((element) -> modelMapper.map(element, RoomDto.class))
@@ -63,6 +80,11 @@ public class RoomServiceImpl implements RoomService{
     public void deleteRoomById(Long roomId) {
         log.info("Deleting the room with ID: {}", roomId);
         Room room=getRoomByID(roomId);
+
+        User user=getCurrentUser();
+        if(!user.equals(room.getHotel().getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+room.getHotel().getId());
+        }
         inventoryService.deleteAllInventories(room)  ;
         roomRepository.delete(room);
     }
@@ -78,4 +100,9 @@ public class RoomServiceImpl implements RoomService{
                 .findById(roomId)
                 .orElseThrow(()-> new ResourceNotFoundException("Room Not Found With ID:"+roomId));
     }
+
+    private User getCurrentUser(){
+        return (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    }
+
 }
